# apps/api-worker/src/routes/configuration.ts
import { Hono } from 'hono';
import { z } from 'zod';
import { zValidator } from '@hono/zod-validator';
import type { AppBindings } from '../env.js';
import { requireRoles, assertGroupAccess } from '../middleware.js';
import { AuditWriter } from '@yrak/audit';

export const configurationRoutes=new Hono<AppBindings>();
const READ=requireRoles('ADMIN','HR','AUDITOR');
const WRITE=requireRoles('ADMIN','HR');
configurationRoutes.get('/groups',READ,async c=>{const u=c.get('user');const rows=await c.env.DB.prepare(`SELECT id,name,description,active FROM groups WHERE organization_id=? ORDER BY name`).bind(u.organizationId).all();return c.json({items:rows.results??[]});});
configurationRoutes.get('/groups/:groupId/levels',READ,async c=>{await assertGroupAccess(c,c.req.param('groupId')!);const rows=await c.env.DB.prepare(`SELECT id,level_number,name,rank_order,active FROM levels WHERE group_id=? ORDER BY rank_order`).bind(c.req.param('groupId')).all();return c.json({items:rows.results??[]});});
configurationRoutes.get('/requirements',READ,async c=>{const u=c.get('user');const rows=await c.env.DB.prepare(`SELECT id,name,requirement_type,validity_days,active,version FROM requirements WHERE organization_id=? ORDER BY name`).bind(u.organizationId).all();return c.json({items:rows.results??[]});});
configurationRoutes.get('/groups/:groupId/rotation-pools',READ,async c=>{await assertGroupAccess(c,c.req.param('groupId')!);const rows=await c.env.DB.prepare(`SELECT p.id,p.source_level_id,p.target_level_id,p.active FROM rotation_pools p WHERE p.group_id=? ORDER BY p.id`).bind(c.req.param('groupId')).all();return c.json({items:rows.results??[]});});
configurationRoutes.get('/rotation-pools/:poolId/queue',READ,async c=>{const u=c.get('user');const rows=await c.env.DB.prepare(`SELECT q.employee_id,e.employee_number,e.name,q.queue_position,q.status,q.last_coverage_at,q.times_selected FROM rotation_queue_entries q JOIN rotation_pools p ON p.id=q.pool_id JOIN employees e ON e.id=q.employee_id WHERE q.pool_id=? AND p.organization_id=? ORDER BY q.queue_position`).bind(c.req.param('poolId'),u.organizationId).all();return c.json({items:rows.results??[]});});
configurationRoutes.get('/users',READ,async c=>{const u=c.get('user');const rows=await c.env.DB.prepare(`SELECT id,email,display_name,role,employee_id,active FROM users WHERE organization_id=? ORDER BY display_name`).bind(u.organizationId).all();return c.json({items:rows.results??[]});});

configurationRoutes.post('/groups',WRITE,zValidator('json',z.object({name:z.string().min(1),description:z.string().optional()})),async c=>{const user=c.get('user'),input=c.req.valid('json'),id=crypto.randomUUID();await c.env.DB.prepare(`INSERT INTO groups(id,organization_id,name,description) VALUES(?,?,?,?)`).bind(id,user.organizationId,input.name,input.description??null).run();await new AuditWriter(c.env.DB).append({organizationId:user.organizationId,actorId:user.id,actorRole:user.role,entityType:'GROUP',entityId:id,action:'CREATED',newValue:input,correlationId:c.get('correlationId')});return c.json({id,...input},201);});
configurationRoutes.post('/groups/:groupId/levels',WRITE,zValidator('json',z.object({number:z.number().int(),name:z.string().min(1),rankOrder:z.number().int()})),async c=>{await assertGroupAccess(c,c.req.param('groupId'));const input=c.req.valid('json'),id=crypto.randomUUID();await c.env.DB.prepare(`INSERT INTO levels(id,group_id,level_number,name,rank_order) VALUES(?,?,?,?,?)`).bind(id,c.req.param('groupId'),input.number,input.name,input.rankOrder).run();return c.json({id,groupId:c.req.param('groupId'),...input},201);});
configurationRoutes.post('/groups/:groupId/transitions',WRITE,zValidator('json',z.object({sourceLevelId:z.string(),targetLevelId:z.string()})),async c=>{await assertGroupAccess(c,c.req.param('groupId'));const input=c.req.valid('json');const levels=await c.env.DB.prepare(`SELECT id FROM levels WHERE group_id=? AND id IN (?,?)`).bind(c.req.param('groupId'),input.sourceLevelId,input.targetLevelId).all();if((levels.results??[]).length!==2)return c.json({error:'LEVEL_GROUP_MISMATCH'},400);const id=crypto.randomUUID();await c.env.DB.prepare(`INSERT INTO level_transitions(id,group_id,source_level_id,target_level_id) VALUES(?,?,?,?)`).bind(id,c.req.param('groupId'),input.sourceLevelId,input.targetLevelId).run();return c.json({id,...input},201);});
configurationRoutes.post('/requirements',WRITE,zValidator('json',z.object({name:z.string(),requirementType:z.enum(['COURSE','CERTIFICATION','PREREQUISITE_EXAM','DOCUMENT','EXPERIENCE','OTHER']),validityDays:z.number().int().positive().nullable().optional()})),async c=>{const input=c.req.valid('json'),user=c.get('user'),id=crypto.randomUUID();await c.env.DB.prepare(`INSERT INTO requirements(id,organization_id,name,requirement_type,validity_days) VALUES(?,?,?,?,?)`).bind(id,user.organizationId,input.name,input.requirementType,input.validityDays??null).run();return c.json({id,...input},201);});
configurationRoutes.put('/levels/:levelId/requirements/:requirementId',WRITE,zValidator('json',z.object({mandatory:z.boolean().default(true),validForEntireCoverage:z.boolean().default(true)})),async c=>{const user=c.get('user'),input=c.req.valid('json');const row=await c.env.DB.prepare(`SELECT l.id,g.id group_id FROM levels l JOIN groups g ON g.id=l.group_id JOIN requirements r ON r.id=? WHERE l.id=? AND g.organization_id=? AND r.organization_id=?`).bind(c.req.param('requirementId'),c.req.param('levelId'),user.organizationId,user.organizationId).first<{id:string;group_id:string}>();if(!row)return c.json({error:'RESOURCE_SCOPE_MISMATCH'},400);await c.env.DB.prepare(`INSERT INTO target_level_requirements(target_level_id,requirement_id,mandatory,valid_for_entire_coverage) VALUES(?,?,?,?) ON CONFLICT(target_level_id,requirement_id) DO UPDATE SET mandatory=excluded.mandatory,valid_for_entire_coverage=excluded.valid_for_entire_coverage`).bind(c.req.param('levelId'),c.req.param('requirementId'),input.mandatory?1:0,input.validForEntireCoverage?1:0).run();return c.json({updated:true});});
configurationRoutes.post('/rotation-pools',WRITE,zValidator('json',z.object({groupId:z.string(),sourceLevelId:z.string(),targetLevelId:z.string(),employeeIds:z.array(z.string()).min(1)})),async c=>{const u=c.get('user'),i=c.req.valid('json');await assertGroupAccess(c,i.groupId);const transition=await c.env.DB.prepare(`SELECT 1 ok FROM level_transitions WHERE group_id=? AND source_level_id=? AND target_level_id=? AND active=1`).bind(i.groupId,i.sourceLevelId,i.targetLevelId).first();if(!transition)return c.json({error:'LEVEL_TRANSITION_NOT_ALLOWED'},400);const employees=await c.env.DB.prepare(`SELECT id FROM employees WHERE organization_id=? AND group_id=? AND base_level_id=? AND id IN (${i.employeeIds.map(()=>'?').join(',')})`).bind(u.organizationId,i.groupId,i.sourceLevelId,...i.employeeIds).all();if((employees.results??[]).length!==i.employeeIds.length)return c.json({error:'ROTATION_EMPLOYEE_SCOPE_MISMATCH'},400);const poolId=crypto.randomUUID();const statements=[c.env.DB.prepare(`INSERT INTO rotation_pools(id,organization_id,group_id,source_level_id,target_level_id) VALUES(?,?,?,?,?)`).bind(poolId,u.organizationId,i.groupId,i.sourceLevelId,i.targetLevelId),...i.employeeIds.map((id,index)=>c.env.DB.prepare(`INSERT INTO rotation_queue_entries(id,pool_id,employee_id,queue_position,status) VALUES(?,?,?,?,'AVAILABLE')`).bind(crypto.randomUUID(),poolId,id,index+1))];await c.env.DB.batch(statements);return c.json({id:poolId,queue:i.employeeIds.map((employeeId,index)=>({employeeId,position:index+1}))},201);});
configurationRoutes.post('/users',WRITE,zValidator('json',z.object({email:z.string().email(),displayName:z.string().min(2),role:z.enum(['ADMIN','HR','SUPERVISOR','COMMITTEE','OPERATOR','EMPLOYEE','AUDITOR']),employeeId:z.string().nullable().optional(),groupIds:z.array(z.string()).default([])})),async c=>{const u=c.get('user'),i=c.req.valid('json'),id=crypto.randomUUID();for(const groupId of i.groupIds)await assertGroupAccess(c,groupId);await c.env.DB.batch([c.env.DB.prepare(`INSERT INTO users(id,organization_id,email,display_name,role,employee_id) VALUES(?,?,?,?,?,?)`).bind(id,u.organizationId,i.email,i.displayName,i.role,i.employeeId??null),...i.groupIds.map(groupId=>c.env.DB.prepare(`INSERT INTO user_groups(user_id,group_id) VALUES(?,?)`).bind(id,groupId))]);return c.json({id,...i},201);});