# openapi/yrak-api.yaml
openapi: 3.1.0
info:
  title: YRAK SUTERM Coberturas API
  version: 0.1.0
  description: >-
    API para coberturas temporales, rotaciones 1–5 días, concursos 6+,
    requisitos, evidencias y auditoría. La IA no tiene autoridad de decisión.
servers:
  - url: https://REPLACE_WITH_API_HOST
security:
  - cloudflareAccess: []
paths:
...
  /v1/config/groups:
    get:
      operationId: listGroups
      responses: { '200': { description: Groups } }
    post:
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses: { '201': { description: Group created } }
  /v1/config/groups/{groupId}/levels:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      operationId: listLevels
      responses: { '200': { description: Levels } }
    post:
      operationId: createLevel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number, name, rankOrder]
              properties:
                number: { type: integer }
                name: { type: string }
                rankOrder: { type: integer }
      responses: { '201': { description: Level created } }
  /v1/config/groups/{groupId}/transitions:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      operationId: createLevelTransition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceLevelId, targetLevelId]
              properties:
                sourceLevelId: { type: string }
                targetLevelId: { type: string }
      responses: { '201': { description: Authorized transition created } }
  /v1/config/requirements:
    get:
      operationId: listRequirements
      responses: { '200': { description: Requirements } }
    post:
      operationId: createRequirement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequirementInput' }
      responses: { '201': { description: Requirement created } }
  /v1/config/levels/{levelId}/requirements/{requirementId}:
    put:
      operationId: attachRequirementToTargetLevel
      parameters:
        - { in: path, name: levelId, required: true, schema: { type: string } }
        - { in: path, name: requirementId, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mandatory: { type: boolean, default: true }
                validForEntireCoverage: { type: boolean, default: true }
      responses: { '200': { description: Mapping updated } }
  /v1/config/rotation-pools:
    post:
      operationId: createRotationPool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupId, sourceLevelId, targetLevelId, employeeIds]
              properties:
                groupId: { type: string }
                sourceLevelId: { type: string }
                targetLevelId: { type: string }
                employeeIds: { type: array, items: { type: string }, minItems: 1 }
      responses: { '201': { description: Rotation queue created } }
  /v1/config/rotation-pools/{poolId}/queue:
    get:
      operationId: listRotationQueue
      parameters:
        - { in: path, name: poolId, required: true, schema: { type: string } }
      responses: { '200': { description: Queue } }
  /v1/config/users:
    get:
      operationId: listUsers
      responses: { '200': { description: Users } }
    post:
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserInput' }
      responses: { '201': { description: User created } }
...
  /v1/calendar/holidays:
    get:
      operationId: listHolidays
      parameters:
        - { name: groupId, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses: { '200': { description: Holidays list } }
    post:
      operationId: createHoliday
      responses: { '201': { description: Holiday created } }
  /v1/calendar/holidays/{holidayId}:
    delete:
      operationId: deleteHoliday
      parameters:
        - { name: holidayId, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: Holiday deleted } }
  /v1/calendar/group-shifts:
    get:
      operationId: listGroupShifts
      parameters:
        - { name: groupId, in: query, required: true, schema: { type: string } }
        - { name: from, in: query, required: true, schema: { type: string } }
        - { name: to, in: query, required: true, schema: { type: string } }
      responses: { '200': { description: Group shift dates list } }
  /v1/calendar/settings/{groupId}:
    get:
      operationId: getCalendarSettings
      parameters:
        - { name: groupId, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: Calendar settings for group } }
    put:
      operationId: updateCalendarSettings
      parameters:
        - { name: groupId, in: path, required: true, schema: { type: string } }
      responses: { '201': { description: Calendar settings versioned } }
...
components:
  securitySchemes:
    cloudflareAccess:
      type: apiKey
      in: cookie
      name: CF_Authorization
      description: Session managed by Cloudflare Access. Development can use x-yrak-user-email only when APP_ENV=development.
  parameters:
    GroupId: { in: path, name: groupId, required: true, schema: { type: string } }
...
  schemas:
...
    RequirementInput:
      type: object
      required: [name, requirementType]
      properties:
        name: { type: string }
        requirementType: { type: string, enum: [COURSE, CERTIFICATION, PREREQUISITE_EXAM, DOCUMENT, EXPERIENCE, OTHER] }
        validityDays: { type: [integer, 'null'] }
...
    UserInput:
      type: object
      required: [email, displayName, role]
      properties:
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [ADMIN, HR, SUPERVISOR, COMMITTEE, OPERATOR, EMPLOYEE, AUDITOR] }
        employeeId: { type: [string, 'null'] }
        groupIds: { type: array, items: { type: string } }